Rearrange types and schema into a new ontology mod
