+++ packages/graph/hash_graph/lib/graph/src/ontology/types/entity_type/mod.rs
@@ -5,14 +5,12 @@ use std::collections::{HashMap, HashSet};
 use error_stack::{ensure, Result};
 use serde::{Deserialize, Serialize};
 
-use crate::types::{
-    schema::{
-        entity_type::links::{Links, ValueOrMaybeOrderedArray},
-        object::{Object, ValidateUri},
-        property_type::PropertyTypeReference,
-        ValidationError, VersionedUri,
-    },
-    BaseUri,
+use crate::ontology::types::{
+    entity_type::links::{Links, ValueOrMaybeOrderedArray},
+    error::ValidationError,
+    property_type::PropertyTypeReference,
+    serde_shared::object::{Object, ValidateUri},
+    uri::{BaseUri, VersionedUri},
 };
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
