+++ packages/graph/hash_graph/lib/graph/src/ontology/types/serde_shared/mod.rs
@@ -1,0 +1,76 @@
+pub mod array;
+pub mod combinator;
+pub mod object;
+
+#[cfg(test)]
+pub mod tests {
+    use std::fmt::Debug;
+
+    use serde::{Deserialize, Serialize};
+
+    /// Will serialize as a constant value `"string"`
+    #[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+    #[serde(rename_all = "camelCase")]
+    pub(in crate::ontology) enum StringTypeTag {
+        #[default]
+        String,
+    }
+
+    #[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+    #[serde(rename_all = "camelCase", deny_unknown_fields)]
+    pub struct StringTypeStruct {
+        r#type: StringTypeTag,
+    }
+
+    pub(in crate::ontology) fn check_serialization<T>(
+        value: &T,
+        json: &serde_json::Value,
+    ) -> Result<(), serde_json::Error>
+    where
+        T: Debug + PartialEq + Serialize,
+    {
+        let serialized_json = serde_json::to_value(value)?;
+        assert_eq!(
+            &serialized_json, json,
+            "Serialized value does not match expected JSON",
+        );
+
+        Ok(())
+    }
+
+    pub(in crate::ontology) fn check_deserialization<T>(
+        value: &T,
+        json: serde_json::Value,
+    ) -> Result<(), serde_json::Error>
+    where
+        for<'de> T: Debug + PartialEq + Deserialize<'de>,
+    {
+        let deserialized_json = serde_json::from_value::<T>(json)?;
+        assert_eq!(
+            &deserialized_json, value,
+            "Deserialized JSON does not match expected value",
+        );
+
+        Ok(())
+    }
+
+    pub(in crate::ontology) fn check<T>(
+        value: &T,
+        json: serde_json::Value,
+    ) -> Result<(), serde_json::Error>
+    where
+        for<'de> T: Debug + PartialEq + Serialize + Deserialize<'de>,
+    {
+        check_serialization(value, &json)?;
+        check_deserialization(value, json)?;
+        Ok(())
+    }
+
+    pub(in crate::ontology) fn check_invalid_json<T>(json: serde_json::Value)
+    where
+        for<'de> T: Debug + Deserialize<'de>,
+    {
+        serde_json::from_value::<T>(json)
+            .expect_err("JSON was expected to be invalid but it was accepted");
+    }
+}
