+++ packages/graph/hash_graph/lib/graph/src/ontology/types/uri.rs
